Added "Save Database" As button
Changed "Save Database" to save the current database without a prompt as expected. "Save Database As" will function how the previous Save button worked.

Added "Database saved" message when the database is successfully saved from one of the save buttons.

Updated the Status displays to include the Status level name that goes with them, allowing the purpose of each level to be more clear. This only applies to the dropdowns.

Increased size of left side item edit/add controls to better fit large content.

Updated the Title Bar to display the current Edit/Add mode of the database.

Fixed the database not properly clearing when a database is currently loaded and an invalid one is attempted to be loaded.